pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "gans_for_maps"
version = "0.0.1"
description = "GANs for generating fantasy maps based on real life satellite pictures"
authors = [
  { name = "ch0fas" },
]
license = { file = "LICENSE" }
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License"
]
dependencies = [
    "loguru",
    "mkdocs",
    "pillow>=12.1.1",
    "pip",
    "pystac-client>=0.9.0",
    "python-dotenv",
    "rasterio>=1.4.4",
    "ruff",
    "s3fs>=2026.3.0",
    "sentinelsat>=1.2.1",
    "torch>=2.5.1",
    "torchvision>=0.20.1",
    "tqdm",
    "typer",
]
requires-python = "~=3.11.0"


[tool.ruff]
line-length = 99
src = ["gans_for_maps"]
include = ["pyproject.toml", "gans_for_maps/**/*.py"]

[tool.ruff.lint]
extend-select = ["I"]  # Add import sorting

[tool.ruff.lint.isort]
known-first-party = ["gans_for_maps"]
force-sort-within-sections = true

